Updated HRPS/src/Control files w/ documentation
Please check the above updates it is important for the Control package. Yesterday, I got a couple of pms from some of you with questions on how some of the methods work. What I noticed was the questions mainly were in the Control package. So proper documentation has been done. Apart from this I finished writing code for 2 classes. Kindly check it as it may be helpful in writing methods for other Control classes as many methods do similar things but with diffrent objects (perhaps). Also regarding the entity classes, I may not document as much as there are primarily getter and setter methods along with some querying methods. If I feel that a method may not be self-explanatory then I will do the documentation. Howewer for the other ones please help me out with the documentation while writing your code. Please reach out in the group or privately if you have any questions.